// trab-web2/src/app/shared/models/via-cep.ts
export class ViaCep {
    cep: string = "";
    logradouro: string = "";
    complemento: string = "";
    unidade: string = "";
    bairro: string = "";
    localidade: string = "";
    uf: string = "";
    estado: string = "";
    regiao: string = "";
    ibge: string = "";
    gia: string = "";
    ddd: string = "";
    siafi: string = "";
    erro?: boolean;
}